move BulkEditColumn-related fields into models/column.py
